Update NameRequestService, extract mapping functions from the service file, so our mapping logic is separate from our service implementation.
