+++ api/namex/resources/name_requests/abstract_nr_resource.py
@@ -52,7 +52,6 @@ class AbstractNameRequestResource(AbstractNROResource, AbstractSolrResource):
 
         return nr_model
 
-
     @staticmethod
     def on_nr_approved(nr, svc):
         """
