+++ api/namex/services/name_request/name_request.py
@@ -2,43 +2,27 @@ from datetime import datetime
 from pytz import timezone
 
 from namex.utils.logging import setup_logging
-from namex.utils.api_resource import log_error
-from namex.utils.common import convert_to_ascii
 
-from namex.constants import NameState, RequestAction, NameRequestActions
+from namex.constants import NameState
 
-from namex.models import Request, Name, State, Comment, Applicant, DecisionReason
+from namex.models import Request, Name, State, Applicant
 
 from .abstract_name_request import AbstractNameRequestMixin
 from .name_request_state import apply_nr_state_change, get_nr_state_actions
 
-from .exceptions import \
-    CreateNameRequestError, SaveNameRequestError, MapRequestDataError, MapRequestHeaderAttributesError, MapRequestAttributesError, \
-    MapRequestApplicantError, MapRequestNamesError, MapPersonCommentError, MapLanguageCommentError, UpdateSubmitCountError, ExtendExpiryDateError
+# Mapping utils used to map HTTP request data to a Request model
+from .mappers.request_header_attrs import map_request_header_attrs
+from .mappers.request_draft_attrs import map_draft_attrs
+from .mappers.request_attrs import map_request_attrs
+from .mappers.request_applicants import map_request_applicant
+from .mappers.request_comments import map_request_comments
+from .mappers.request_names import map_submitted_name
+
+from .utils import get_item_from_list
+from .exceptions import CreateNameRequestError, SaveNameRequestError, MapRequestDataError, \
+    MapRequestApplicantError, MapRequestNamesError, UpdateSubmitCountError, ExtendExpiryDateError
 
 setup_logging()  # Important to do this first
-
-NAME_REQUEST_SOURCE = 'NAMEREQUEST'
-
-
-def build_language_comment(english_bol, user_id, nr_id):
-    lang_comment = Comment()
-    lang_comment.examinerId = user_id
-    lang_comment.nrId = nr_id
-    if english_bol is True:
-        # Add a comment for the examiner that says this is not an english name
-        lang_comment.comment = 'The applicant has indicated the submitted name or names are in English.'
-    else:
-        lang_comment.comment = 'The applicant has indicated the submitted name or names are not English.'
-    return lang_comment
-
-
-def build_name_comment(user_id, nr_id):
-    name_comment = Comment()
-    name_comment.examinerId = user_id
-    name_comment.nrId = nr_id
-    name_comment.comment = 'The submitted name or names is a person name, coined phrase or trademark'
-    return name_comment
 
 
 class NameRequestService(AbstractNameRequestMixin):
@@ -100,6 +84,10 @@ class NameRequestService(AbstractNameRequestMixin):
     def virtual_wc_service(self, service):
         self._virtual_wc_service = service
 
+    @property
+    def new_state_code(self):
+        return self.next_state_code if self.next_state_code else self.request_state_code
+
     def create_name_request(self):
         """
         # !Important! All new name requests should be initially set to the DRAFT state.
@@ -138,20 +126,7 @@ class NameRequestService(AbstractNameRequestMixin):
 
     @classmethod
     def get_item_from_list(cls, items, item_id, item_prop='id'):
-        """
-        TODO: We could make a util for this...
-        :param items:
-        :param item_id:
-        :param item_prop:
-        :return:
-        """
-        matches = [i for i in items if i.__getattribute__(item_prop) == item_id]
-        if len(matches) == 0:
-            return None
-        if len(matches) == 1:
-            return matches[0]
-        if len(matches) > 1:
-            raise Exception('More than one match found for a given ID!')
+        return get_item_from_list(items, item_id, item_prop=item_prop)
 
     @classmethod
     def update_request_submit_count(cls, name_request):
@@ -180,196 +155,55 @@ class NameRequestService(AbstractNameRequestMixin):
 
         return name_request
 
-    # Methods used to map request data
-    def map_request_data(self, name_request, map_draft_attrs=False):
+    def map_request_data(self, name_request, map_attrs=False):
         """
         This method maps data from the HTTP request data over to the name request.
         We use this to set draft attributes, header attributes, and comments...
         :param name_request:
-        :param map_draft_attrs:
+        :param map_attrs:
         :return:
         """
-        new_state_code = self.next_state_code if self.next_state_code else self.request_state_code
-
         # Set the request attributes
-        name_request = self.map_request_attrs(name_request)
+        name_request = map_request_attrs(
+            name_request,
+            nr_id=self.nr_id,
+            nr_num=self.nr_num,
+            request_entity=self.request_entity,
+            request_action=self.request_action,
+            request_type=self.request_type
+        )
 
-        # If this is a draft, set name request header attributes
-        if map_draft_attrs:
-            name_request = self.map_draft_attrs(name_request)
+        # If this is a DRAFT, set draft attributes
+        if map_attrs:
+            name_request = map_draft_attrs(name_request, user_id=self.user_id)
 
-        name_request = self.map_request_header_attrs(name_request)
-        name_request = self.map_request_comments(name_request)
+        # Map request header attributes
+        name_request = map_request_header_attrs(
+            name_request,
+            request_data=self.request_data,
+            user_id=self.user_id
+        )
+
+        # Map request comments
+        name_request = map_request_comments(
+            name_request,
+            request_data=self.request_data,
+            user_id=self.user_id,
+            nr_id=self.nr_id
+        )
 
         try:
-            if new_state_code == State.COND_RESERVE:
+            if self.new_state_code == State.COND_RESERVE:
                 name_request.consentFlag = 'Y'
 
-            if new_state_code in [State.RESERVED, State.COND_RESERVE]:
-                    name_request.expirationDate = self.create_expiry_date(
-                        start=name_request.submittedDate,
-                        expires_in_days=56,
-                        tz=timezone('UTC')
+            if self.new_state_code in [State.RESERVED, State.COND_RESERVE]:
+                name_request.expirationDate = self.create_expiry_date(
+                    start=name_request.submittedDate,
+                    expires_in_days=56,
+                    tz=timezone('UTC')
                 )
         except Exception as err:
             raise MapRequestDataError(err)
-
-        return name_request
-
-    def map_draft_attrs(self, name_request):
-        """
-        Used internally by map_request_data.
-        :param name_request:
-        :return:
-        """
-        try:
-            user_id = self.user_id
-            # Set this to name_request_service_account
-            name_request.userId = user_id
-            name_request.submittedDate = datetime.utcnow()
-        except Exception as err:
-            raise MapRequestDataError(err)
-
-        return name_request
-
-    def map_request_attrs(self, name_request):
-        """
-        Used internally by map_request_data.
-        :param name_request:
-        :return:
-        """
-        try:
-            # Use class property values for the ID, NR Number and Source!
-            # Do not map those values from the request if supplied, as they
-            # should not be changed outside of the context of this application!
-            nr_id = self.nr_id
-            nr_num = self.nr_num
-            name_request.id = nr_id
-            name_request.nrNum = nr_num
-            name_request._source = NAME_REQUEST_SOURCE
-
-            # Default to whatever entity, action, or type already exists when mapping
-            request_entity = self.request_entity if self.request_entity else name_request.entity_type_cd
-            request_action = self.request_action if self.request_action else name_request.request_action_cd
-            request_type = self.request_type if self.request_type or self.request_type is None else name_request.requestTypeCd
-
-            # Set action and entity
-            if request_entity:
-                name_request.entity_type_cd = request_entity
-
-            if request_action:
-                name_request.request_action_cd = request_action
-
-            # TODO: Throw exceptions for invalid combos?
-            if not request_type and request_entity and request_action:
-                # If request_type is None (eg. no 'requestTypeCd' was provided in the payload)
-                # but a request_entity (entity_type_cd) and a request_action (request_action_cd)
-                # are supplied, use get_mapped_request_type to map the requestTypeCd in the model
-                # using the action and entity type
-                request_type = self.get_mapped_request_type(request_entity, request_action)
-                name_request.requestTypeCd = request_type[0]
-            elif request_type is not None:
-                # If request_type is NOT None, (eg. 'requestTypeCd' was provided in the payload)
-                # then use the provided value
-                name_request.requestTypeCd = request_type
-        except Exception as err:
-            raise MapRequestAttributesError(err)
-
-        return name_request
-
-    def map_request_header_attrs(self, name_request):
-        """
-        Used internally by map_request_data.
-        :param name_request:
-        :return:
-        """
-        user_id = self.user_id
-        request_data = self.request_data
-
-        try:
-            # TODO: Review additional info stuff from NRO/namex (prev NR for re-applies,no NWPTA?
-            if isinstance(request_data.get('natureBusinessInfo'), str):
-                name_request.natureBusinessInfo = convert_to_ascii(request_data.get('natureBusinessInfo'))
-            if isinstance(request_data.get('additionalInfo'), str):
-                name_request.additionalInfo = convert_to_ascii(request_data.get('additionalInfo'))
-            if isinstance(request_data.get('tradeMark'), str):
-                name_request.tradeMark = request_data.get('tradeMark')
-            if isinstance(request_data.get('previousRequestId'), int):
-                name_request.previousRequestId = request_data.get('previousRequestId')
-            if isinstance(request_data.get('priorityCd'), str):
-                name_request.priorityCd = request_data.get('priorityCd')
-            if request_data.get('priorityCd') == 'Y':
-                name_request.priorityDate = datetime.utcnow().date()
-
-            name_request.submitter_userid = user_id
-
-            # XPRO
-            if isinstance(request_data.get('xproJurisdiction'), str):
-                name_request.xproJurisdiction = request_data.get('xproJurisdiction')
-            # For MRAS participants
-            if isinstance(request_data.get('homeJurisNum'), str):
-                name_request.homeJurisNum = convert_to_ascii(request_data.get('homeJurisNum'))
-            # For existing businesses
-            if isinstance(request_data.get('corpNum'), str):
-                name_request.corpNum = convert_to_ascii(request_data.get('corpNum'))
-        except Exception as err:
-            raise MapRequestHeaderAttributesError(err)
-
-        return name_request
-
-    def map_request_comments(self, name_request):
-        """
-        Used internally by map_request_data. Execute any logic required to map comments here.
-        :param name_request:
-        :return:
-        """
-        name_request = self.map_request_language_comments(name_request)
-        name_request = self.map_request_person_name_comments(name_request)
-
-        return name_request
-
-    def map_request_language_comments(self, name_request):
-        """
-        Used internally by map_request_comments.
-        :param name_request:
-        :return:
-        """
-        try:
-            request_data = self.request_data
-            user_id = self.user_id
-            nr_id = self.nr_id
-
-            # If the language comment exists, we don't need to add it again, after nitail PIST it gets set to None so must dealt with
-            if request_data.get('english') is not None:
-                lang_comment = build_language_comment(request_data.get('english'), user_id, nr_id)
-                matching_comments = list(filter(lambda x: x.comment == lang_comment.comment, list(name_request.comments)))
-                if len(matching_comments) == 0:
-                    name_request.comments.append(lang_comment)
-        except Exception as err:
-            raise MapLanguageCommentError(err)
-
-        return name_request
-
-    def map_request_person_name_comments(self, name_request):
-        """
-        Used internally by map_request_comments.
-        :param name_request:
-        :return:
-        """
-        try:
-            request_data = self.request_data
-            user_id = self.user_id
-            nr_id = self.nr_id
-
-            if request_data.get('nameFlag') is not None:
-                if request_data.get('nameFlag') is True:
-                    # If the person name comment exists, we don't need to add it again
-                    name_comment = build_name_comment(user_id, nr_id)
-                    matching_comments = list(filter(lambda x: x.comment == name_comment.comment, list(name_request.comments)))
-                    if len(matching_comments) == 0:
-                        name_request.comments.append(name_comment)
-        except Exception as err:
-            raise MapPersonCommentError(err)
 
         return name_request
 
@@ -379,47 +213,39 @@ class NameRequestService(AbstractNameRequestMixin):
         :param name_request:
         :return:
         """
-        request_data = self.request_data
-        nr_id = self.nr_id
-
-        request_applicants = request_data.get('applicants')
-        applicants = []
-
         try:
+            request_data = self.request_data
+            nr_id = self.nr_id
+
+            request_applicants = request_data.get('applicants')
+            applicants = []
+
+            def create_or_update_applicant(mapped_applicants, applicant_data):
+                if request_applicant_id:
+                    existing_applicants = name_request.applicants.all()
+                    match = get_item_from_list(existing_applicants, request_applicant_id, 'partyId')
+                    if match:
+                        applicant = map_request_applicant(match, applicant_data)
+                        mapped_applicants.append(applicant)
+                else:
+                    applicant = self.create_applicant()
+                    applicant.nrId = nr_id
+
+                    applicant = map_request_applicant(applicant, applicant_data)
+                    mapped_applicants.append(applicant)
+
+                return applicants
+
             if isinstance(request_applicants, list):
                 for request_applicant in request_applicants:
                     request_applicant_id = request_applicant.get('partyId')
-
-                    if request_applicant_id:
-                        existing_applicants = name_request.applicants.all()
-                        match = self.get_item_from_list(existing_applicants, request_applicant_id, 'partyId')
-                        if match:
-                            applicant = self.map_request_applicant(match, request_applicant)
-                            applicants.append(applicant)
-                    else:
-                        applicant = self.create_applicant()
-                        applicant.nrId = nr_id
-
-                        applicant = self.map_request_applicant(applicant, request_applicant)
-                        applicants.append(applicant)
+                    applicants = create_or_update_applicant(applicants, request_applicant)
 
                 name_request.applicants = applicants
 
             elif isinstance(request_applicants, dict):
                 request_applicant_id = request_applicants.get('partyId')
-
-                if request_applicant_id:
-                    existing_applicants = name_request.applicants.all()
-                    match = self.get_item_from_list(existing_applicants, request_applicant_id, 'partyId')
-                    if match:
-                        applicant = self.map_request_applicant(match, request_applicants)
-                        applicants.append(applicant)
-                else:
-                    applicant = self.create_applicant()
-                    applicant.nrId = nr_id
-
-                    applicant = self.map_request_applicant(applicant, request_applicants)
-                    applicants.append(applicant)
+                applicants = create_or_update_applicant(applicants, request_applicants)
 
                 name_request.applicants = applicants
 
@@ -427,33 +253,6 @@ class NameRequestService(AbstractNameRequestMixin):
             raise MapRequestApplicantError(err)
 
         return name_request
-
-    def map_request_applicant(self, applicant, request_applicant):
-        applicant.lastName = convert_to_ascii(request_applicant['lastName'])
-        applicant.firstName = convert_to_ascii(request_applicant['firstName'])
-        if request_applicant['middleName']:
-            applicant.middleName = convert_to_ascii(request_applicant['middleName'])
-        applicant.contact = convert_to_ascii(request_applicant['contact'])
-        if request_applicant['clientFirstName']:
-            applicant.clientFirstName = convert_to_ascii(request_applicant['clientFirstName'])
-        if request_applicant['clientLastName']:
-            applicant.clientLastName = convert_to_ascii(request_applicant['clientLastName'])
-        if request_applicant['phoneNumber']:
-            applicant.phoneNumber = convert_to_ascii(request_applicant['phoneNumber'])
-        if request_applicant['faxNumber']:
-            applicant.faxNumber = convert_to_ascii(request_applicant['faxNumber'])
-        applicant.emailAddress = convert_to_ascii(request_applicant['emailAddress'])
-        applicant.addrLine1 = convert_to_ascii(request_applicant['addrLine1'])
-        if request_applicant['addrLine2']:
-            applicant.addrLine2 = convert_to_ascii(request_applicant['addrLine2'])
-        if request_applicant['addrLine3']:
-            applicant.addrLine3 = convert_to_ascii(request_applicant['addrLine3'])
-        applicant.city = convert_to_ascii(request_applicant['city'])
-        applicant.stateProvinceCd = request_applicant['stateProvinceCd']
-        applicant.postalCd = convert_to_ascii(request_applicant['postalCd'])
-        applicant.countryTypeCd = request_applicant['countryTypeCd']
-
-        return applicant
 
     def map_request_names(self, name_request):
         """
@@ -472,181 +271,33 @@ class NameRequestService(AbstractNameRequestMixin):
                     match = self.get_item_from_list(existing_names, request_name_id)
                     if match:
                         # Update the name
-                        updated_name = self.map_submitted_name(match, request_name)
+                        updated_name = map_submitted_name(
+                            match,
+                            request_name,
+                            nr_id=self.nr_id,
+                            new_state_code=self.new_state_code,
+                            request_entity=self.request_entity,
+                            request_action=self.request_action
+                        )
+
                         name_request.names.append(updated_name)
                 else:
                     submitted_name = self.create_name()
-                    submitted_name = self.map_submitted_name(submitted_name, request_name)
+                    submitted_name = map_submitted_name(
+                        submitted_name,
+                        request_name,
+                        nr_id=self.nr_id,
+                        new_state_code=self.new_state_code,
+                        request_entity=self.request_entity,
+                        request_action=self.request_action
+                    )
+
                     name_request.names.append(submitted_name)
 
         except Exception as err:
             raise MapRequestNamesError(err)
 
         return name_request
-
-    def map_submitted_name(self, submitted_name, name):
-        """
-        Used internally by map_request_names.
-        :param submitted_name:
-        :param name:
-        :return:
-        """
-        new_state_code = self.next_state_code if self.next_state_code else self.request_state_code
-
-        # Common name attributes
-        submitted_name = self.map_submitted_name_attrs(submitted_name, name)
-        test_conflict = name.get('conflict1')
-        if len(test_conflict) > 0:
-            conflict_flag = 'Y'
-        else:
-            conflict_flag = 'N'
-
-        if new_state_code in [State.COND_RESERVE] and conflict_flag == 'Y':
-            submitted_name = self.map_submitted_name_conflicts(submitted_name, name)
-
-        consent_words_list = name.get('consent_words', None)
-        if consent_words_list and len(consent_words_list) > 0:
-            submitted_name = self.map_submitted_name_consent_words(submitted_name, consent_words_list)
-
-        ## add macros for specific entities and actions
-        if new_state_code in [State.COND_RESERVE, State.RESERVED]:
-            macro_list = []
-            if self.request_action == 'MVE' and self.request_entity != 'UL':
-                macro_list.append('Continuation')
-            if self.request_action == 'MVE' and self.request_entity == 'UL':
-                macro_list.append('Ulc Cont In')
-            if len(macro_list) > 0 :
-                submitted_name = self.map_submitted_name_macros(submitted_name, macro_list)
-
-        return submitted_name
-
-    def map_submitted_name_attrs(self, submitted_name, name):
-        """
-        Used internally by map_submitted_name.
-        :param submitted_name:
-        :param name:
-        :return:
-        """
-        new_state_code = self.next_state_code if self.next_state_code else self.request_state_code
-
-        try:
-            submitted_name.nrId = self.nr_id
-            submitted_name.choice = name.get('choice', 1)
-            submitted_name.name_type_cd = name.get('name_type_cd', 'CO')
-            submitted_name.name = convert_to_ascii(name.get('name', ''))
-            submitted_name.designation = name.get('designation', '')
-            # For existing businesses
-            if isinstance(name.get('corpNum'), str):
-                # To clear the corpNum use an empty string in the data payload
-                submitted_name.corpNum = convert_to_ascii(name.get('corpNum'))
-
-            if new_state_code == State.DRAFT:
-                submitted_name.state = NameState.NOT_EXAMINED.value
-
-            elif new_state_code == State.COND_RESERVE:
-                submitted_name.state = NameState.COND_RESERVE.value
-
-            elif new_state_code == State.RESERVED:
-                submitted_name.state = NameState.RESERVED.value
-
-            elif new_state_code == State.CONDITIONAL:
-                submitted_name.state = NameState.CONDITION.value
-
-            elif new_state_code == State.APPROVED:
-                submitted_name.state = NameState.APPROVED.value
-
-        except Exception as err:
-            raise MapRequestNamesError(err, 'Error setting common name attributes.')
-
-        return submitted_name
-
-    def map_submitted_name_conflicts(self, submitted_name, name):
-        """
-        Used internally by map_submitted_name.
-        :param submitted_name:
-        :param name:
-        :return:
-        """
-        try:
-            # Only capturing one conflict
-            if name.get('conflict1_num'):
-                submitted_name.conflict1_num = name.get('conflict1_num')
-            if name.get('conflict1'):
-                submitted_name.conflict1 = name.get('conflict1')
-            # Conflict text same as Namex
-            submitted_name.decision_text = 'Consent is required from ' + name.get('conflict1') + '\n' + '\n'
-        except Exception as err:
-            raise MapRequestNamesError(err, 'Error on reserved conflict info.')
-
-        return submitted_name
-
-    def map_submitted_name_macros(self, submitted_name,macro_list):
-        for macro in macro_list:
-            try:
-                macro_text = None
-                if macro != '' or len(macro) > 0:
-                    macro_text = DecisionReason.find_by_name(macro)
-            except Exception as err:
-                log_error('Error on get decision reason word. Macro Word[0]'.format(macro), err)
-                raise MapRequestNamesError('Error mapping macro words.')
-
-            try:
-                if submitted_name.decision_text is None:
-                    submitted_name.decision_text = macro + '- ' + macro_text.reason + '\n' + '\n'
-                else:
-                    submitted_name.decision_text += macro + '- ' + macro_text.reason + '\n' + '\n'
-
-            except Exception as err:
-                raise MapRequestNamesError(err, 'Error adding macro words to decision.')
-
-
-
-        return submitted_name
-
-
-    def clear_submitted_name_conflicts(self, submitted_name):
-        """
-        Used internally by map_submitted_name.
-        :param submitted_name:
-        :param name:
-        :return:
-        """
-        try:
-            submitted_name.conflict1_num = None
-            submitted_name.conflict1 = None
-        except Exception as err:
-            raise MapRequestNamesError(err, 'Error on draft empty conflict info.')
-
-        return submitted_name
-
-    def map_submitted_name_consent_words(self, submitted_name, consent_list):
-        """
-        Used internally by map_submitted_name.
-        :param submitted_name:
-        :param consent_list:
-        :return:
-        """
-        decision_text = submitted_name.decision_text
-        for consent in consent_list:
-            try:
-                cnd_instructions = None
-                if consent != '' or len(consent) > 0:
-                    cnd_instructions = self.virtual_wc_service.get_word_condition_instructions(consent)
-            except Exception as err:
-                log_error('Error on get consent word. Consent Word[0]'.format(consent), err)
-                raise MapRequestNamesError('Error mapping consent words.')
-
-            try:
-                if decision_text is None:
-                    decision_text = consent + '- ' + cnd_instructions + '\n' + '\n'
-                else:
-                    decision_text += consent + '- ' + cnd_instructions + '\n' + '\n'
-
-                submitted_name.decision_text = decision_text
-            except Exception as err:
-                raise MapRequestNamesError(err, 'Error adding consent words to decision.')
-
-        return submitted_name
 
     def apply_state_change(self, name_request, next_state, on_success=None):
         """
